$().bind is deprecated in favor of $().on
"As of jQuery 1.7, the .on() method is the preferred method for attaching event handlers to a document." I'm using `.on()` to observe clicks on the whole document, and then select only the ones that hit `a.page-scroll`. That's because elements could be rendered after this is called, so delegated events are equivalent to the way .bind() worked before.